CHROMIUM: REVERT ME: Turn on panel in dark resume
Due to a complicated set of inter-dependencies, keeping the panel off in
dark resume causes a catastrophic memory leak in chrome's gpu process,
which eventually results in chrome invoking the oom_killer followed soon
after by a system crash.

While several solutions and workarounds exist to deal with this, the
most complete solution involves making the gpu process aware of dark
resume so that it doesn't attempt to touch the display when the system
is in that state.

Implementing this will take some time so as a short-term solution, allow
the panel to turn on in dark resume.  We still don't want the screen to
flash however so don't allow the panel to turn on the backlight.  The
downside to this patch is that the system wastes about 20 milliwatts in
S3 so this patch should be reverted as soon as the long-term solution is
ready.

BUG=chrome-os-partner:35699
TEST=manual on samus
activate_short_dark_resume
powerd_dbus_suspend

Wait for a minute or two and observe that the screen does not flash.
Resume the system and check /var/log/ui/ui.LATEST to confirm that there
are no error messages about SwapBuffers failing.  Additionally, run
'ps aux --sort -%mem' after resume and verify that the gpu process is
consuming a reasonable amount of memory.

Change-Id: I3c1810309265f99520694e4f4cefeeb6bf61abb1
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/245873
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 1c58d0238f17c64947d4c76d8232ee5412f18867)
Reviewed-on: https://chromium-review.googlesource.com/246240