CHROMIUM: rockchip: fix dmcclk race
We need to kick the cpus at the start of setting the dmcclk rate. We pause all
of the other cpus with smp_call_function and don't wait for the cpus to complete
the IPI. If we timeout on the last clk_set_rate and call clk_set_rate fast
enough, a cpu may still have a pending IPI when we set the boolean to keep the
cpus paused. At this point, we will get a soft lockup when we call
smp_call_function.

BUG=chrome-os-partner:31440
TEST=boot jerry

Change-Id: I83c7ab36167fe630cd62c620ff01367b3c8fddc8
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/246701
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
(cherry picked from commit 1bf9443b0a841ab9e0739224d4768ff9e06633a7)
Reviewed-on: https://chromium-review.googlesource.com/251458