CHROMIUM: iwl7000: don't try to stop scans that are not running anymore
In certain conditions, mac80211 may ask us to stop a scan (scheduled
or normal) that is not running anymore.  This can also happen when we
are doing a different type of scan, for instance, mac80211 can ask us
to stop a scheduled scan when we are running a normal scan, due to
some race conditions.  In this case, we would stop the wrong type of
scan and leave everything everything in a wrong state.

To fix this, simply ignore scan stop requests for scans types that are
not running.

BUG=chrome-os-partner:36253
TEST=verify scan works after "start sched scan, start scan, stop sched
     scan" with a sleep in sched_scan_stopped_work() to trigger the race

Change-Id: I3f765a316ae947af448606b0b37903e52f429449
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/246893
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277595
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Grace Kao <grace.kao@intel.com>
Tested-by: Grace Kao <grace.kao@intel.com>