UPSTREAM: HID: rmi: Print the firmware id of the touchpad
Knowing the firmware id is extremely useful when debugging issues related to
the touchpad. It can be used to determine the hardware, firmware version,
and configuation of the touchpad. This patch queries the firmware id and
prints it as the touchpad is starting so that it will show up in the dmesg
output included in bug reports.

BUG=chrome-os-partner:39003
TEST=Firmware ID is printed on Jerry.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 70e003f75fab6c1496147838acf59be85d2520c0)
Signed-off-by: Andrew Duggan <aduggan@synaptics.com>

Change-Id: I1c2db6a14f9f3b070ba0924f775c337c21675554
Reviewed-on: https://chromium-review.googlesource.com/265089
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Andrew Duggan <aduggan@synaptics.com>
Commit-Queue: Andrew Duggan <aduggan@synaptics.com>
Reviewed-on: https://chromium-review.googlesource.com/270057