CHROMIUM: rockchip: dmc: calculate timeout earlier
This calculates the timeout for dmc set rate earlier by passing in a ktime_t
pointer to rockchip_dmc_wait. This is set in the vop_isr function so there's no
chance for an interrupt or softirq to preempt us before we calculate the
timeout. This also moves to ktime_t from a u64 representation of nanoseconds.
The vblank time is calculated ahead of time in mode_set. To make sure we don't
use an incorrect value for this, we lock the dmc to prevent a dmc clk rate
change from happening while adjusting the mode.

BUG=chrome-os-partner:31440
TEST=boot jerry

Change-Id: I7768d352ba28e56c40ba3bd896b57aa3ad7391f5
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/248941
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 257c4e601aad6f64195f7dfcb012f8b18630c6db)
Reviewed-on: https://chromium-review.googlesource.com/251464