CHROMIUM: iwl7000 mvm: Fix ROC removal
iwl_mvm_stop_roc removes TE only if running flag is set. This is not
correct since this flag is only set when the TE is started.
This resulted in a TE not being removed, when mac80211 believes that
there are no active ROCs.

BUG=N/A
TEST=cancel a remain-on-channel request immediately after issuing

Change-Id: I28303496a3b4fcc4a3dfef1d540f971131f1545b
Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/256044
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277601
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Grace Kao <grace.kao@intel.com>
Tested-by: Grace Kao <grace.kao@intel.com>