FROMLIST: ath10k: fix quiet duration configuration
Duty cycle is nothing but percentage of one period in which a
signal is active. But it is wrongly interpreted as ratio of quiet
duration. So fix the quiet duration to the remaining percentage
of duty cycle (active ratio).

Signed-off-by: Rajkumar Manoharan <rmanohar@qti.qualcomm.com>

TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Change-Id: Id3a647d8273710486d180eac9e16b9c6d81ac23a
Reviewed-on: https://chromium-review.googlesource.com/253498
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Tested-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Toshi Kikuchi <toshik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256030
Reviewed-by: Patrick Sosinski <sosinski@google.com>
Commit-Queue: Patrick Sosinski <sosinski@google.com>
Tested-by: Patrick Sosinski <sosinski@google.com>