CHROMIUM: rockchip: disable dmc freq with short vblank
This disables dmc freq when the vblank time is too short for a display mode. It
keeps track of whether rockchip_dmc_disable was already called. This calculates
the minimum time for vblank with the max time for setting the dmc rate in sram
plus the minimum time to pause the cpus. This is because we can timeout while
pausing the cpus, but not while setting the rate in sram.

BUG=chrome-os-partner:37086
TEST=boot minnie, see that dmc doesn't time out constantly in dmesg

Change-Id: Ia547b52c7a0478de6be995b42d5c3bc68a880163
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/253893
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 35787df54592ffb28e44c21ee0b6d055720b081b)
Reviewed-on: https://chromium-review.googlesource.com/259720
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>