FROMLIST: ath10k: Make HTT fill size configurable
The HTT RX ring is replenished with a maximum of 16 buffers,
but this might be insufficient when RX traffic is high.
Not having enough RX buffers throttles the FW, resulting
in low throughput.

This patch adds a module parameter to adjust the fill size
based on the platform/usage.

Signed-off-by: Sujith Manoharan <c_manoha@qca.qualcomm.com>

TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Change-Id: I7e283ad2d404a25ab49d732da114288cb1217c0b
Reviewed-on: https://chromium-review.googlesource.com/253504
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Tested-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Toshi Kikuchi <toshik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256036
Reviewed-by: Patrick Sosinski <sosinski@google.com>
Commit-Queue: Patrick Sosinski <sosinski@google.com>
Tested-by: Patrick Sosinski <sosinski@google.com>