CHROMIUM: hid: rmi: Disable scanning if the device is not a wake source
When the lid is open and the system goes into S3 the touchpad firmware will
remain scanning at a minimal rate allowing a finger touch to wake the system.
But, when the lid is closed scanning should be disabled to save power. This
CL sends the commands to the touchpad to disable power when it is not a wake
source.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>

BUG=chrome-os-partner:39003
TEST=Touchpad did not scan when the wakeup sysfs file was set to disabled.

Change-Id: I776f46d7a84000c5e6fbf4df4123b57665aababd
Reviewed-on: https://chromium-review.googlesource.com/265281
Tested-by: Andrew Duggan <aduggan@synaptics.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Andrew Duggan <aduggan@synaptics.com>
Reviewed-on: https://chromium-review.googlesource.com/272465