UPSTREAM: thermal: cpu_cooling: Remove cpu_dev update on policy CPU update
It was initially understood that an update to the cpu_device
(cached in cpufreq_cooling_device) was required to ascertain the
correct operating point of the device on a cpufreq policy->cpu update
or creation or deletion of a cpufreq policy.
(e.g. when the existing policy CPU goes offline).

This update is not required and it is possible to ascertain the OPPs
from the leading CPU in a cpufreq domain even if the CPU is hotplugged out.

Fixes: e0128d8ab423 ("thermal: cpu_cooling: implement the power cooling device API")
Acked-by: Javi Merino <javi.merino@arm.com>
Signed-off-by: Kapileshwar Singh <kapileshwar.singh@arm.com>
Signed-off-by: Eduardo Valentin <edubezval@gmail.com>
(cherry picked from commit 54b92aae834ce8d02d6092f8ae5eb1fe0d607f1e)
Signed-off-by: Kevin Hilman <khilman@linaro.org>

BUG=chrome-os-partner:42804
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: I52376cb35123373a627c15805b8412b1d29832b6
Signed-off-by: Dawei.Chien <dawei.chien@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/283835
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/286440
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>