CHROMIUM: X86: Explicitly add -fomit-frame-pointer to boot/compressed.
We are changing the compiler default in ChromiumOS to fno-omit-frame-pointer.
The makefile under boot/compressed does not specify fomit-frame-pointer
directly so it gets the compiler default, which at the moment is
fomit-frame-pointer. Since we are going to change the default, we will
add the option explicitly to the Makefile. Not doing that causes the
kernel image to exceed its size limits.

BUG=chromium:464650
TEST=cbuildbot --hwtest lumpy-release
Signed-off-by: Luis A Lozano <llozano@chromium.org>

Change-Id: Ic7df21f83152f7a9145e408cfc79f928b7b8ba50
Reviewed-on: https://chromium-review.googlesource.com/260445
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>