FROMLIST: mmc: dw_mmc: Don't send clock off command if power is off
We've already seen that it causes failures if we try to send the
"SDMMC_CMD_UPD_CLK" command when power is off to the system.  Here's
another case where we were doing it: we were sending an
SDMMC_CMD_UPD_CLK to set the clock to 0 when the card was already
powered off and the clock was already at 0.  Let's avoid that command
and avoid problems.

BUG=chrome-os-partner:36163, chrome-os-partner:37702
TEST=Reboot tests

Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6059011/)
Change-Id: I66ed57b54b0d8f9071ee1a1f03f23c5d8daf488b
Reviewed-on: https://chromium-review.googlesource.com/260340
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 68136398112894eb79b9c62e300abe9760cb06a7)
Reviewed-on: https://chromium-review.googlesource.com/261847