mfd: cros ec: spi: Increase wait time to 200ms
This is a sucky change to bump up the time we'll wait for the EC.  Why
is it sucky?  If 200ms for a transfer is a common thing it will have a
massively bad impact on keyboard responsiveness.

It still seems like a good idea to do this, though, because we have a
gas gauge that claims that in an extreme case it could stretch the i2c
clock for 144ms.  It's not a common case so it shouldn't affect
responsiveness, but it can happen.  It's much better to have a single
slow keyboard response than to start returning errors when we don't
have to.

In newer EC designs we should probably implement a virtual battery to
respond to the kernel to insulate the kernel from these types of
issues.

BUG=chrome-os-partner:37960
TEST=Normal transfers still work OK; put bogus delay in EC and no errors.

Original-Change-Id: I2af07dbccfb12f4c136c5c9011cccb64b7e9f90b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261896
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit 7fce72005d61d2383e576baf7305942b2d84a99a)

Change-Id: Iaa3fc842b4e5cb9f39cbb14fbc017c797ddec8b3
Reviewed-on: https://chromium-review.googlesource.com/262350
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>