CHROMIUM: dts: rk3288-veyron: Add SPI2 rx-sample-delay-ns
We found that our SPI lines can have too high capacitance for 49.5MHz
operation when a Servo cable is plugged in. Counteract this with an
(empirically determined) Rx Sample Delay.

BUG=chrome-os-partner:38339
TEST=Ran flashrom on Jerry, confirmed that it worked at 49.5MHz and
that the RSD value in CTRLR0 was '1' when booted with a 99MHz parent
clock and '2' when booted with a 198MHz parent clock.

Change-Id: Ia99ad1470e78497a493848de80e3e955f5cd0ad2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262469
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit f350c3376587d31b6d89d5c85eadecf09308c848)
Reviewed-on: https://chromium-review.googlesource.com/263143
Tested-by: Douglas Anderson <dianders@chromium.org>