ARM: dts: Document max98090 interrupt disabling on rk3288-veyron as a hack
In the change (b2dd0d1 CHROMIUM: ASOC: max98090: remove unused the
interrupt pin of codec) we took out the pinctrl statement for the
interrupt of the max98090 but we didn't actually remove the interrupt,
apparently because the driver requires it.

The net effect of removing the pinctrl statement is to leave the pin
in the default boot state, which is a pulldown.  That means we've got
a pulldown on an active low signal, effectively disabling it.

Let's at least document our HACK so we don't confuse the next person
that deals with it (like maybe someone trying to upstream things).
Ideally we could fix the driver to make this interrupt optional in the
very least, even if we don't want to debug the hang that was observed.

BUG=chrome-os-partner:38518
TEST=Still see no interrupts for max98090_interrupt in /proc/interrupts

Change-Id: I5039583e51eb30749706fb2287b05514bd98946b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/263240
Reviewed-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit f215c76001b489720a1f5769e7d553404379e5b0)
Reviewed-on: https://chromium-review.googlesource.com/271999