CHROMIUM: hid: rmi: Set power state on inhibit/uninhibit callbacks
Save power when the device is inhibited and restore power when it is not
inhibited. This CL sends the power commands to disable scanning on inhibit
and return to full power on uninhibit.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>

BUG=chrome-os-partner:39003
TEST=Ensured that the appropriate commands were sent to the touchpad when the
touchpad was inhibited and uninhibited.

Change-Id: Id0964003f14b43771dafc10a1efa0702b005a625
Reviewed-on: https://chromium-review.googlesource.com/265282
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Andrew Duggan <aduggan@synaptics.com>
Tested-by: Andrew Duggan <aduggan@synaptics.com>
(cherry picked from commit b8878547ec8e01fa2cc625f475cf7408559c0659)
Reviewed-on: https://chromium-review.googlesource.com/270547
Reviewed-by: Benson Leung <bleung@chromium.org>