BACKPORT: clk: tegra: Implement SOR1 clock on Tegra210
SOR1 clock uses a two-stage mux to select a parent. This is difficult to
represent in software, especially if drivers need to select one of the
parents during the programming of the SOR module.

To model the two-stage mux using a single level of parents, use a parent
and mask/value mapping.

Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
[seanpaul fixed minor comment collision]
(cherry picked from commit 117dd69616ab0cc816828b7e225851b7e6ead19e)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Iea89c491974a95deb4464256458db97e21926467
Reviewed-on: https://chromium-review.googlesource.com/285868
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>