UPSTREAM: clk: tegra: Implement two-stage mux clock
Some clocks on Tegra have a two-stage mux that is difficult to represent
in the common clock framework. While this could be modeled using several
layers of parents, that becomes cumbersome since some drivers need to
select among the set of parents during programming of their hardware
module. It is easier to model them as a single hierarchy of parents and
use a mask/value to parent mapping to represent the relationship.

Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
(cherry picked from commit 20ca91f7ca44a60ef0f869792a7997bb67715f8d)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Iee07586828d4bedc29ba6efd8cf34419c0d2b5cc
Reviewed-on: https://chromium-review.googlesource.com/285867
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>