Revert "mmc: core: Hold a wake lock accross delayed work + mmc rescan"
This reverts commit 08054072af22f5220571e4408ca27ea8e96a33e1.

In the CL (0805407 mmc: core: Hold a wake lock accross delayed work +
mmc rescan) a wake lock was added to make sure that we could detect an
SD card that was inserted right around the time the system was going
into suspend (or maybe runtime suspend?).

That CL really should have added a wake_lock per MMC host, not a
single wake_lock for the whole MMC core subsystem.  Without adding a
wakelock per host the different MMC users (eMMC, SD, SDIO-WiFi) will
clobber each other's requests.

The fact that there was _not_ one wake_lock per MMC host actually hid
a bug.  mmc_rescan() has a few exit points that don't release the
wake_lock, the most notable one being for NONREMOVABLE cards.

This bug shows the race on rk3288-veyron systems.  On those systems we
have 3 users of MMC (enumerated in this order):
- mmc1: SD card   - removable
- mmc0: eMMC      - non-removable
- mmc2: SDIO-WIFI - non-removable

If we add a enough delay before calling the wake_lock() for mmc2 then
the mmc_rescan() for mmc1 will have time to fully run and have time to
either call wake_lock_timeout() or wake_unlock() _before_ the mmc2
request for wake_lock().  This shows that for mmc2 nobody was ever
releasing the wake lock.  This is demonstrated in
<http://crosreview.com/270185>.

It wouldn't be terribly hard to move the wake_lock to be per host and
to make sure we clear the wake_lock at all of the exit points of
mmc_rescan(), but I'd rather someone who understood the motivation of
the original CL do that (the original CL has no description of why it
was needed nor links to any bugs).  For now, a revert seems safest
without a good way for me to test the problem that the original CL was
fixing.

BUG=chrome-os-partner:39991
TEST=suspend_stress_test

Change-Id: I15081bf677adfe2aeff0302bc85636df3ebc0887
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270186
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
(cherry picked from commit d479d781a2161a08a99ceaf74e5b2c4b3c87a3d6)
Reviewed-on: https://chromium-review.googlesource.com/270371