UPSTREAM: thermal: support slope and offset coefficients
It is common to have a linear extrapolation from
the current sensor readings and the actual temperature
value. This is specially the case when the sensor
is in use to extrapolate hotspots.

This patch adds slope and offset constants for
single sensor linear extrapolation equation. Because
the same sensor can be use in different locations,
from board to board, these constants are added
as part of thermal_zone_params.

The constants are available through sysfs.

It is up to the device driver to determine
the usage of these values.

Signed-off-by: Eduardo Valentin <edubezval@gmail.com>
(cherry picked from commit 9d0be7f4810257a9b0fc78fff641f14409f14ab3)
Signed-off-by: Kevin Hilman <khilman@linaro.org>

BUG=chrome-os-partner:42804
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: I7ded9c9bee1b35d16b7d6798f96c11abc1a54e1b
Signed-off-by: Dawei.Chien <dawei.chien@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/283838
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/286443
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>