CHROMIUM: rockchip: dmc: split idling of domains
This splits the idling of the core domain from the rest of the domains. This
is so the other domains can be idled before turning off the MMU. It also
splits the deidling of the domains so memory can be restored when multi-rank
training is done in a following CL before deidling the noncore domains.

BUG=chrome-os-partner:38282
TEST=boot jaq

Change-Id: I6df55882d46b44395eff624a38a0d8b84c4cedfb
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270555
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
(cherry picked from commit 91d15803f10a2d1bfa81fcba3f175da77233117e)
Reviewed-on: https://chromium-review.googlesource.com/289243
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>