CHROMIUM: WORKAROUND: xhci: Resume hub as long as as not suspending.
Root cause of this issue is that XHCI controller is creating an
infinite supply of interrupts once an DP alternate mode device has
been connected to the port.

This patch enables resuming the root hub providing we're not
suspending as that act cleanses whatever part of the controller is
generating the interrupts.

This patch should NOT be carried forward.

BUG=chrome-os-partner:39105,chrome-os-partner:35320
TEST=manual

  1. boot samus without any external USB devices, login & get shell
  2. from shell,
     watch -n 1 grep acpi /proc/interrupts
  3. connect hoho || dingdong to one of the type-C ports
  4. No longer see interrupt storm on acpi interrupt.

Also verify that:

  1. XHCI controller is still entering runtime suspend via
     watch -n 1 cat \
       /sys/bus/usb/devices/usb*/power/runtime_suspended_time
  2. usb 3.0 & usb 2.0 devices still enumerate successfully.

Change-Id: I4650ccebe7e427fa5ea3f2fe1bc72025a1efed13
Signed-off-by: Todd Broch <tbroch@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270803
Reviewed-by: Sameer Nanda <snanda@chromium.org>