CHROMIUM: rockchip: dmc: train all ranks
This performs data training on all ranks instead of just rank0 when setting
the rate in the dmc. It saves and restores the memory that is overwritten
due to data training for all ranks instead of allocating a large buffer.
This is due to the interleaving of ranks being too large to allocate in a
buffer.

BUG=chrome-os-partner:38282
TEST=boot speedy

Change-Id: I4854ceae22a065b6cb6ed9fb02ad035c039c7986
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/270806
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
(cherry picked from commit bc15e7cf64d1153dcd2f2389237e98d9fb8a3b0c)
Reviewed-on: https://chromium-review.googlesource.com/289217
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>