CHROMIUM: dmcfreq: update cpufreq directly
This updates the cpufreq directly in the devfreq driver instead of updating the
cpufreq policy. It looks like updating the policy was racing with the cpu
cooling, so just prevent any other cpufreq changes then set it to max.

This moves the calls of cpu_hotplug_disable/enable so the policy semaphore is
not held when those are called to avoid possible deadlocks.

BUG=chrome-os-partner:39636
TEST=boot jerry

Change-Id: Ic4ba48f2150c6204c8e3599f78edd274bf31c31e
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276261
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 23628863dea819bfe183feda5dc10cec1a8045dc)
Reviewed-on: https://chromium-review.googlesource.com/278182
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>