HACK: ARM: dts: Pretend bt_host_wake is a wakeup key
Here's how we ought to do Bluetooth wakeup:

1. If we are doing Bluetooth over SDIO then wakeup ought to be part of
the SDIO subsystem and ought to be a shared signal for Bluetooth and
WiFi (we use the same transport for both, so why would we use
different wakeup signals?).  See http://crosreview.com/276878 for a
proposal for how to do this.

2. If we are doing Bluetooth over UART, we should have some sort of
handshaking signal above and beyond the standard CTS / RTS signals.
CTS and RTS prevent the other side from sending you data.  Without
considering power consumption, it's expected that CTS / RTS are
deasserted only temporarily due to a FIFO full condition, so no other
lines are needed.  However, once we start using CTS/RTS for other
reasons (preventing sending data while we are asleep) we need a 2nd
set of pins to request that the other side should wake up.  This is
sometimes talked about as LPM (low power mode?).  Once we have LPM
implmented then the bt_host_wake should be handled by that, but as a
stopgap we might be able to use this patch.

BUG=chrome-os-partner:41503
TEST=On Broadcom can wake from S3 via Bluetooth keyboard

Change-Id: Ia21a0f20e2dd4bb16804499aed87ef7fb2a6229e
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/278190
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit d57df798c2d3e88b046acff6a23a5d33855f1655)
Reviewed-on: https://chromium-review.googlesource.com/281813