CHROMIUM: iwl7000: mvm: work around SG/CSUM link on old kernels
On older kernels, NETIF_F_SG requires at least one CSUM offload and
will be disabled if no CSUM offload is present. This causes a warning
as the disabling happens while the netdev isn't yet operating.

For the short term, work around this by not advertising SG in this
case. As it's probably possible to trigger the same warning as the
user with ethtool, we need to find another way to avoid it as well.

Reported-by: Beni Lev <beni.lev@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: ef94affca48be3eae5a2169bf91a73a22c35c762