Revert "CHROMIUM: clk/tegra: Initialize disp emc clocks to 656MHz"
This increments the reference count of disp{1,2}.emc so that they
may never be disabled, even if they are unused (as is the case for
disp2).  This effectively pins the EMC clock to 1.6Ghz.

For now we can rely on the EMC clock being 800Mhz at boot, which
should be sufficient until the DC driver requests its bandwidth.
Ideally, the DC driver should request an initial bandwidth before
drawing the first frame.

This reverts commit d36a1d14a09a88fb3f39d36df20aabefdd80f885.

BUG=chrome-os-partner:41674
TEST=Build and boot on Smaug, still no DC underflows.  EMC now
scales below 1.6Ghz at runtime.

Change-Id: I41a7161aa0639fbb3f90b1eba82629f364af5139
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285448
Reviewed-by: Sean Paul <seanpaul@google.com>
Reviewed-on: https://chromium-review.googlesource.com/286229
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>