CHROMIUM: drm/tegra: Do not round bandwidth to EMC rate
There's no need to round the requested bandwidth to an EMC rate
for bandwidth clients - the EMC driver will handle calculating
an appropriate rate based on the bandwidth requested.

Also, do not skip the clk_set_rate() if the requested bandwidth is 0.
The DC's bandwidth request should get cleared if the display is off.

BUG=chrome-os-partner:41674
TEST=Build and boot on Smaug.  EMC rate drops when display turns off.

Change-Id: Ia2465541a182676e589bb56d5cda0feea8c784b0
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285449
Reviewed-by: Sean Paul <seanpaul@google.com>
Reviewed-on: https://chromium-review.googlesource.com/286350
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>