drm/nouveau: mmu: Invalidate L2 after modifying PDE/PTE tables
Although the L2 is yet enabled for GPU pages, but the PTE and PDE tables
are cached. So it's necessary to invalidate the L2 after altering the
tables to keep the GPU view on them coherent.

To achieve this, the patch extends the nvkm_interface with flush and
invalidate methods and forks the gk104 and gm107 ltc implementations
into gk20a and gm20b to implement them for appropriate GPUs.

Change-Id: Ibeb3ff6c3e27d113b0a41dd222798fa35d5514b5
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/287280
Reviewed-by: Anatol Pomazau <anatol@google.com>
Tested-by: Anatol Pomazau <anatol@google.com>