UPSTREAM: drm/tegra: sor: Hook up I2C-over-AUX to output
This is necessary for the output abstraction to retrieve a list of valid
modes from the EDID of a connected panel/monitor. This will be useful in
conjunction with DisplayPort support that will be added in a subsequent
patch, so that the driver can read EDID via the AUX channel.

Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
(cherry picked from commit d5198802278ee7b80ac3ea4c15b80b6c0ec1d1dc)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Ia88af2bf89650c6ddbb1fb015fc720915eb4019e
Original-Change-Id: I85aa2679ff9e75bbc0d2cf84d8d05b068c6da886
Reviewed-on: https://chromium-review.googlesource.com/298348
Reviewed-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>