CHROMIUM: rockchip: dmc: Avoid glitches due to slow CPU
When trying to thermal throttle the DMC we found glitches on the
screen.  We tracked that down to the fact that when the CPU was
running slow we weren't giving enough time for DMC to change
frequencies.

This attempts to solve things in two ways:
* If the CPU is < 1.2GHz we don't try to do any DMC changes.
* If the CPU is between 1.2 and 1.4 GHz we increase the time.

BUG=chrome-os-partner:42854
TEST=No more glitching when running hot.

Change-Id: I231df0ac9952b79f00220fc0293cc826ea2cdbaa
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289595
Tested-by: Lin Huang <hl@rock-chips.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
(cherry picked from commit 74a76ac56c7bc11eac377d5a5ea8dbc2fd842c20)
Reviewed-on: https://chromium-review.googlesource.com/294791